extract code responsible for view canvas
