+++ pytripgui/controller/ctx.py
@@ -1,13 +1,10 @@
-import logging
-# import matplotlib.colors
-
-# import numpy as np
 import matplotlib.pyplot as plt
 
+import logging
 logger = logging.getLogger(__name__)
 
 
-class Ctx(object):
+class Ctxx(object):
     """
     This class holds logic for plotting CTX stuff.
     """
@@ -29,7 +26,7 @@ class Ctx(object):
         ctx = plc._model.plot.ctx
         pm = plc._model.plot
 
-        figure = plc.figure
+        # figure = plc.figure
         axes = plc.axes
 
         if ctx is None:
@@ -83,13 +80,14 @@ class Ctx(object):
         # strictly the contrast bar should be set in the viewer?
         # setup the HU bar:
         if not plc.hu_bar:
-            cax = figure.add_axes([0.1, 0.1, 0.03, 0.8])
-            cb = figure.colorbar(plc.axim_ctx, cax=cax)
-            cb.set_label("HU", color=pm.fg_color, fontsize=pm.cb_fontsize)
-            cb.outline.set_edgecolor(pm.bg_color)
-            cb.ax.yaxis.set_tick_params(color=pm.fg_color, labelsize=pm.cb_fontsize)
-            plt.setp(plt.getp(cb.ax.axes, 'yticklabels'), color=pm.fg_color)
-            plc.hu_bar = cb
+            pass
+            # cax = figure.add_axes([0.1, 0.1, 0.03, 0.8])
+            # cb = figure.colorbar(plc.axim_ctx, cax=cax)
+            # cb.set_label("HU", color=pm.fg_color, fontsize=pm.cb_fontsize)
+            # cb.outline.set_edgecolor(pm.bg_color)
+            # cb.ax.yaxis.set_tick_params(color=pm.fg_color, labelsize=pm.cb_fontsize)
+            # plt.setp(plt.getp(cb.ax.axes, 'yticklabels'), color=pm.fg_color)
+            # plc.hu_bar = cb
 
     @staticmethod
     def change_contrast(plc, contrast):
