+++ pytripgui/view/viewcanvas_widget.py
@@ -10,7 +10,7 @@ logger = logging.getLogger(__name__)
 
 class ViewCanvasWidget(FigureCanvas):
 
-    def __init__(self, width=6, height=4, dpi=110):
+    def __init__(self, width=6, height=4, dpi=110, bg_color = 'black'):
         self.fig = Figure(figsize=(width, height), dpi=dpi)
 
         # Here one can adjust the position of the CTX plot area.
@@ -24,3 +24,6 @@ class ViewCanvasWidget(FigureCanvas):
         # next too lines are needed in order to catch keypress events in plot canvas by mpl_connect()
         FigureCanvas.setFocusPolicy(self, QtCore.Qt.ClickFocus)
         FigureCanvas.setFocus(self)
+
+        # initial setup of the ViewCanvas
+        self.fig.patch.set_facecolor(bg_color)
